# backend-api/openapi/openapi.yaml
components:
...
  schemas:
...
    DtoAuditRecord:
      properties:
        date:
          format: date-time
          type: string
        idAuditRecord:
          format: int32
          type: integer
        module:
          type: string
        operationDetail:
          type: string
        users:
          $ref: '#/components/schemas/DtoUser'
        valido:
          type: boolean
      type: object
    DtoBudget:
      properties:
        date:
          format: date-time
          type: string
        encabezado:
          type: string
        idBudget:
          format: int32
          type: integer
        items:
          items:
            $ref: '#/components/schemas/DtoItem'
          type: array
        listaProcedures:
          items:
            $ref: '#/components/schemas/DtoProcedure'
          type: array
        notes:
          type: string
        number:
          format: int32
          type: integer
        person:
          $ref: '#/components/schemas/DtoPerson'
        propertyAmount:
          type: number
        saldo:
          deprecated: true
          type: number
        status:
          type: string
        total:
          deprecated: true
          type: number
        valido:
          type: boolean
        version:
          format: int32
          type: integer
      type: object
...
    DtoDeed:
      properties:
        body:
          type: string
        dateDeedrecording:
          format: date-time
          type: string
        dateRegistration:
          format: date-time
          type: string
        folios:
          items:
            $ref: '#/components/schemas/DtoFolio'
          type: array
        idDeed:
          format: int32
          type: integer
        notes:
          type: string
        number:
          format: int32
          type: integer
        procedures:
          items:
            $ref: '#/components/schemas/DtoProcedure'
          type: array
        registrationEntryNumber:
          type: string
        status:
          type: string
        testimonios:
          items:
            $ref: '#/components/schemas/DtoTestimony'
          type: array
        valido:
          type: boolean
        version:
          format: int32
          type: integer
      type: object
    DtoDeedManagement:
      properties:
        clientReferencia:
          $ref: '#/components/schemas/DtoPerson'
        dateStart:
          format: date-time
          type: string
        encabezado:
          type: string
        idManagement:
          format: int32
          type: integer
        listaClientesInvolucrados:
          items:
            $ref: '#/components/schemas/DtoPerson'
          type: array
        listaProceduresAsociados:
          items:
            $ref: '#/components/schemas/DtoProcedure'
          type: array
        notes:
          type: string
        number:
          format: int32
          type: integer
        numberArchivo:
          format: int32
          type: integer
        numberBibliorato:
          format: int32
          type: integer
        personNotary:
          $ref: '#/components/schemas/DtoPerson'
        recordHistory:
          items:
            $ref: '#/components/schemas/DtoHistory'
          type: array
        status:
          $ref: '#/components/schemas/DtoManagementStatus'
        valido:
          type: boolean
        version:
          format: int32
          type: integer
      type: object
...
    DtoDocumentType:
      properties:
        deliveredBy:
          type: string
        dueDays:
          format: int32
          type: integer
        enabled:
          type: boolean
        expires:
          type: boolean
        idDocumentType:
          format: int32
          type: integer
        name:
          type: string
        returned:
          type: boolean
        templateProcedures:
          items:
            $ref: '#/components/schemas/DtoProcedureTemplate'
          type: array
        templateTramiteses:
          items:
            $ref: '#/components/schemas/DtoProcedureTemplate'
          type: array
          writeOnly: true
        valido:
          type: boolean
        version:
          format: int32
          type: integer
      type: object
...
    DtoHistory:
      properties:
        date:
          format: date-time
          type: string
        estadosDeManagement:
          $ref: '#/components/schemas/DtoManagementStatus'
        gestionesDeEscrituras:
          $ref: '#/components/schemas/DtoDeedManagement'
        idHistory:
          format: int32
          type: integer
        notes:
          type: string
        valido:
          type: boolean
        version:
          format: int32
          type: integer
      type: object
...
    DtoIdentificationType:
      properties:
        characters:
          type: string
        idIdentificationType:
          format: int32
          type: integer
        name:
          type: string
        valido:
          type: boolean
        version:
          format: int32
          type: integer
      type: object
...
    DtoManagementStatus:
      properties:
        data:
          additionalProperties: {}
          type: object
        historials:
          items: {}
          type: array
          uniqueItems: true
        idManagementStatus:
          format: int32
          type: integer
        name:
          type: string
        notes:
          type: string
        valido:
          type: boolean
        version:
          format: int32
          type: integer
      type: object
...
    DtoPerson:
      properties:
        address:
          type: string
        birthDate:
          format: date-time
          type: string
        client:
          type: boolean
        copiases:
          items: {}
          type: array
          uniqueItems: true
        dtoIdentificationType:
          $ref: '#/components/schemas/DtoIdentificationType'
        email:
          type: string
        firstName:
          type: string
        folioses:
          items: {}
          type: array
          uniqueItems: true
        id:
          format: int32
          type: integer
        identificacioneses:
          items: {}
          type: array
          uniqueItems: true
        identificationNumber:
          type: string
        isClient:
          type: boolean
        lastName:
          type: string
        listDtoManagementDeEscriturasDeNotary:
          items:
            $ref: '#/components/schemas/DtoDeedManagement'
          type: array
        listaDtoManagementDeEscriturasPerson:
          items:
            $ref: '#/components/schemas/DtoDeedManagement'
          type: array
        listaProceduresPerson:
          items:
            $ref: '#/components/schemas/DtoProcedure'
          type: array
        maritalStatus:
          type: string
        marriageCount:
          format: int32
          type: integer
        nationality:
          type: string
        notaryRegistrationNumber:
          format: int32
          type: integer
        occupation:
          type: string
        phone:
          type: string
        presupuestoses:
          items: {}
          type: array
          uniqueItems: true
        sex:
          type: string
        suplenciasesForFkIdSubstitute:
          items: {}
          type: array
          uniqueItems: true
        suplenciasesForFkIdSubstituted:
          items: {}
          type: array
          uniqueItems: true
        taxId:
          type: string
        usuarioses:
          items: {}
          type: array
          uniqueItems: true
        version:
          format: int32
          type: integer
      type: object
    DtoProcedure:
      properties:
        budget:
          $ref: '#/components/schemas/DtoBudget'
        deed:
          $ref: '#/components/schemas/DtoDeed'
        deedManagement:
          $ref: '#/components/schemas/DtoDeedManagement'
          writeOnly: true
        idProcedure:
          format: int32
          type: integer
        listaDocumentsManagement:
          items:
            $ref: '#/components/schemas/DtoSubmittedDocument'
          type: array
        listaDocumentsNecesarios:
          items:
            $ref: '#/components/schemas/DtoDocumentType'
          type: array
        listaDocumentsNoPrecentados:
          items:
            $ref: '#/components/schemas/DtoSubmittedDocument'
          type: array
        listaDocumentsPresentados:
          items:
            $ref: '#/components/schemas/DtoSubmittedDocument'
          type: array
        listaPersons:
          items:
            $ref: '#/components/schemas/DtoPerson'
          type: array
        listaPresupuestos:
          items:
            $ref: '#/components/schemas/DtoBudget'
          type: array
        management:
          $ref: '#/components/schemas/DtoDeedManagement'
        managementDeEscrituras:
          $ref: '#/components/schemas/DtoDeedManagement'
        notes:
          type: string
        procedureType:
          $ref: '#/components/schemas/DtoProcedureType'
        property:
          $ref: '#/components/schemas/DtoProperty'
        tiposDeProcedure:
          $ref: '#/components/schemas/DtoProcedureType'
          writeOnly: true
        valido:
          type: boolean
      type: object
...
    DtoUser:
      properties:
        idUser:
          format: int32
          type: integer
        name:
          type: string
        password:
          type: string
        persons:
          $ref: '#/components/schemas/DtoPerson'
        roleId:
          format: int32
          type: integer
        roleName:
          type: string
        status:
          type: boolean
        token:
          type: string
        type:
          type: string
        valido:
          type: boolean
        version:
          format: int32
          type: integer
      type: object
...
    DtoWorkflowNode:
      properties:
        id:
          format: int32
          type: integer
        positionX:
          format: float
          type: number
        positionY:
          format: float
          type: number
        statusManagementId:
          format: int32
          type: integer
        statusManagementName:
          type: string
        type:
          type: string
        version:
          format: int32
          type: integer
        workflowDefinitionId:
          format: int32
          type: integer
      type: object
    DtoWorkflowTransition:
      properties:
        condition:
          type: string
        description:
          type: string
        destinationNodeId:
          format: int32
          type: integer
        id:
          format: int32
          type: integer
        originNodeId:
          format: int32
          type: integer
        version:
          format: int32
          type: integer
        workflowDefinitionId:
          format: int32
          type: integer
      type: object
...
    PersonInfo:
      properties:
        idPerson:
          format: int32
          type: integer
        lastName:
          type: string
        name:
          type: string
      type: object
...
    RoleInfo:
      properties:
        idRole:
          format: int32
          type: integer
        name:
          type: string
      type: object
    RoleRequest:
      properties:
        active:
          type: boolean
        description:
          type: string
        modulos:
          items:
            type: string
          type: array
        name:
          minLength: 1
          type: string
      required:
      - name
      type: object
    RoleResponse:
      properties:
        active:
          type: boolean
        description:
          type: string
        idRole:
          format: int32
          type: integer
        modulos:
          items:
            type: string
          type: array
        name:
          type: string
      type: object
...
    UserRequest:
      properties:
        active:
          type: boolean
        name:
          minLength: 1
          type: string
        password:
          type: string
        type:
          minLength: 1
          type: string
      required:
      - name
      - type
      type: object
    UserResponse:
      properties:
        active:
          type: boolean
        idUser:
          format: int32
          type: integer
        name:
          type: string
        person:
          $ref: '#/components/schemas/PersonInfo'
        role:
          $ref: '#/components/schemas/RoleInfo'
        type:
          type: string
      type: object
...
info:
  contact:
    email: info@licensis.com
    name: Licensis
    url: https://www.licensis.com
  description: Sistema de gestión notarial — API REST backend. Todos los recursos siguen el prefijo `/api/v1/`. Las respuestas
    de error siguen el esquema `{error, message, timestamp, path}`.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Notaire API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /api/v1/audit-log/user/{idUser}:
    get:
      operationId: getByUser
      parameters:
      - in: path
        name: idUser
        required: true
        schema:
          format: int32
          type: integer
      responses:
        '200':
          content:
            '*/*':
              schema:
                items:
                  $ref: '#/components/schemas/DtoAuditRecord'
                type: array
          description: OK
      summary: Obtener registros de auditoria por usuario
      tags:
      - RegistroAuditoria
...
  /api/v1/roles:
    get:
      operationId: getAllRoles
      responses:
        '200':
          content:
            '*/*':
              schema:
                items:
                  $ref: '#/components/schemas/RoleResponse'
                type: array
          description: OK
      summary: Obtener todos los roles
      tags:
      - Roles
    post:
      operationId: createRole
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleRequest'
        required: true
      responses:
        '201':
          content:
            '*/*':
              schema:
                type: object
          description: Creado
        '400':
          content:
            '*/*':
              schema:
                type: object
          description: Solicitud inválida
        '409':
          content:
            '*/*':
              schema:
                type: object
          description: Conflicto
      summary: Crear nuevo rol
      tags:
      - Roles
  /api/v1/roles/usuarios/{idUser}:
    delete:
      operationId: unassignRoleFromUser
      parameters:
      - in: path
        name: idUser
        required: true
        schema:
          format: int32
          type: integer
      responses:
        '204':
          description: Rol desasignado
        '404':
          description: Usuario no encontrado
      summary: Desasignar rol de usuario
      tags:
      - Roles
  /api/v1/roles/{idRole}/usuarios/{idUser}:
    put:
      operationId: assignRoleToUser
      parameters:
      - in: path
        name: idRole
        required: true
        schema:
          format: int32
          type: integer
      - in: path
        name: idUser
        required: true
        schema:
          format: int32
          type: integer
      responses:
        '200':
          content:
            '*/*':
              schema:
                type: object
          description: OK
      summary: Asignar rol a usuario
      tags:
      - Roles
  /api/v1/roles/{id}:
    delete:
      operationId: deleteRole
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: int32
          type: integer
      responses:
        '204':
          description: Eliminado
        '404':
          description: No encontrado
      summary: Eliminar rol
      tags:
      - Roles
    get:
      operationId: getRoleById
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: int32
          type: integer
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RoleResponse'
          description: OK
        '404':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RoleResponse'
          description: No encontrado
      summary: Obtener rol por ID
      tags:
      - Roles
    put:
      operationId: updateRole
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: int32
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleRequest'
        required: true
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RoleResponse'
          description: OK
        '404':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RoleResponse'
          description: No encontrado
      summary: Actualizar rol
      tags:
      - Roles
...
  /api/v1/usuarios/{id}:
    delete:
      operationId: deleteUser
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: int32
          type: integer
      responses:
        '204':
          description: Eliminado
        '404':
          description: No encontrado
      summary: Eliminar usuario
      tags:
      - Usuarios
    get:
      operationId: getUserById
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: int32
          type: integer
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserResponse'
          description: OK
        '404':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserResponse'
          description: No encontrado
      summary: Obtener usuario por ID
      tags:
      - Usuarios
    put:
      operationId: updateUser
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: int32
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
        required: true
      responses:
        '200':
          description: OK
        '404':
          description: No encontrado
      summary: Actualizar usuario
      tags:
      - Usuarios
...
  /api/v1/workflow-node:
    post:
      operationId: create_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DtoWorkflowNode'
        required: true
      responses:
        '201':
          content:
            '*/*':
              schema:
                type: object
          description: Creado
        '400':
          content:
            '*/*':
              schema:
                type: object
          description: Solicitud inválida
        '409':
          content:
            '*/*':
              schema:
                type: object
          description: Conflicto
      summary: Crear nodo en un workflow
      tags:
      - Workflow Node
  /api/v1/workflow-node/by-workflow/{workflowId}:
    get:
      operationId: getByWorkflow_1
      parameters:
      - in: path
        name: workflowId
        required: true
        schema:
          format: int32
          type: integer
      responses:
        '200':
          content:
            '*/*':
              schema:
                items:
                  $ref: '#/components/schemas/DtoWorkflowNode'
                type: array
          description: OK
      summary: Obtener nodos por workflow
      tags:
      - Workflow Node
  /api/v1/workflow-node/{id}:
    delete:
      operationId: delete_1
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: int32
          type: integer
      responses:
        '204':
          content:
            '*/*':
              schema:
                type: object
          description: Eliminado
        '404':
          content:
            '*/*':
              schema:
                type: object
          description: No encontrado
      summary: Eliminar nodo
      tags:
      - Workflow Node
    get:
      operationId: getById_1
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: int32
          type: integer
      responses:
        '200':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DtoWorkflowNode'
          description: OK
        '404':
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DtoWorkflowNode'
          description: No encontrado
      summary: Obtener nodo por ID
      tags:
      - Workflow Node
    put:
      operationId: update_1
      parameters:
      - in: path
        name: id
        required: true
        schema:
          format: int32
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DtoWorkflowNode'
        required: true
      responses:
        '200':
          content:
            '*/*':
              schema:
                type: object
          description: OK
        '404':
          content:
            '*/*':
              schema:
                type: object
          description: No encontrado
      summary: Actualizar nodo
      tags:
      - Workflow Node
...
  /api/v1/workflow-transition/by-workflow/{workflowId}:
    get:
      operationId: getByWorkflow
      parameters:
      - in: path
        name: workflowId
        required: true
        schema:
          format: int32
          type: integer
      responses:
        '200':
          content:
            '*/*':
              schema:
                items:
                  $ref: '#/components/schemas/DtoWorkflowTransition'
                type: array
          description: OK
      summary: Obtener transiciones por workflow
      tags:
      - Workflow Transition
...
servers:
- description: Desarrollo local
  url: http://localhost:8080
- description: Docker local
  url: http://localhost:8081
...